Prevent deletion of dev builds

+++ Jenkinsfile
@@ -3,11 +3,12 @@ node {
     def COMMIT = "${git.GIT_COMMIT}"
     def BRANCH = "${git.GIT_BRANCH}"
 
+    def IS_MASTER = "${BRANCH}" == "master"
 
     def devApp
 
     stage('Build dev') {
-        def devTag = ("${BRANCH}" == "master") ? "dev" : "${git.GIT_BRANCH}"
+        def devTag = IS_MASTER ? "dev" : "${git.GIT_BRANCH}"
         devApp = docker.build("shiro/home-fe:${devTag}", "--rm -f docker/fe/Dockerfile .")
     }
 
@@ -20,12 +21,14 @@ node {
             }
         }
     } finally {
-        echo 'Removing docker image'
-        sh "docker rmi ${devApp.id}"
+        if(!IS_MASTER) {
+            echo 'Removing docker image'
+            sh "docker rmi ${devApp.id}"
+        }
     }
 
 
-    if ("${BRANCH}" == "master") {
+    if (IS_MASTER) {
         def prodApp
 
         stage('Build prod') {
